Auth protection for Post Controller

+++ app/Http/Controllers/PostController.php
@@ -34,18 +34,27 @@ class PostController extends Controller
 
     public function getAdminIndex()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $posts = Post::orderBy('created_at', 'desc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
 
     public function getAdminCreate()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
 
     public function getAdminEdit($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $post = Post::find($id);
         $tags = Tag::all();
         return view('admin.edit', ['post' => $post, 'tags' => $tags]);
@@ -53,6 +62,10 @@ class PostController extends Controller
 
     public function postAdminCreate(Request $request)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
+
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10',
@@ -76,6 +89,10 @@ class PostController extends Controller
 
     public function postAdminUpdate(Request $request)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
+
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10',
@@ -94,6 +111,10 @@ class PostController extends Controller
 
     public function getAdminDelete($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
+
         $post = Post::find($id);
         $post->likes()->delete();
         $post->tags()->detach();
